improved jitter - added color change when "match" is completed
A moving average is used to moderate the jitter. It may be set with the
4th line of the configuration file - vowelShapeConfig.txt. The tolerance
for the match is specified on the 5th line. The Graph "match" parameter
is the absolute distance between the active and match point. The
Triangle and Oval match is a hack of the ratio of the areas and ratio of
the bx/bys.