Final Changes - added back input devices - draw timing changes
Added back the input device menu as a trial feature - not fully vetted.
Moved the undraw and draw commands together to see if timing issues can
me mediated on slow machines.
Adding know issue to the Readme and adding the Vowel_Shapes_about.pdf to
the repository
Addind cxfreeze.txt with the command line to produce an .exe